+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.5.0
+- The API endpoint can now be configured through the Raygun client
+- Fix: A bad access exception when logging response codes from the Raygun API
+
 ## 1.4.2
 - Updated crash report property name for the stack trace
 - Additional guards against the provider generating exceptions